+++ OpenGPSLogger/app/src/main/java/eu/tijlb/opengpslogger/ui/view/bitmap/LastLocationBitmapRenderer.kt
@@ -26,7 +26,7 @@ class LastLocationBitmapRenderer(val context: Context) : AbstractBitmapRenderer(
 
     override suspend fun draw(
         bbox: BBoxDto,
-        zoom: Int,
+        zoom: Double,
         renderDimension: Pair<Int, Int>,
         assignBitmap: (Bitmap) -> Unit,
         refreshView: () -> Any
